Created the RwOps system.
Added a memory buffer to represent void* that are a space in memory.

Added custom marshalers for the RwOps and MemoryBuffer classes.

Added the SDL allocation methods for the RwOps class.

Added the RwOps class to represent the native RwOps structure.

Added the RwOps types.

Added the internal native structure.

Added the SDL_free() functionality.